requirements.txt
clickclick>=0.9
keyring
PyYAML
requests
stups-zign>=0.9
